+++ pkg/util/k8sutil/names/id.go
@@ -30,7 +30,15 @@ import (
 )
 
 func GetArangodID(group api.ServerGroup) string {
-	return fmt.Sprintf("%s%s", GetArangodIDPrefix(group), strings.ToLower(uniuri.NewLen(8)))
+	return GetArangodIDPredefined(group, strings.ToLower(uniuri.NewLen(8)))
+}
+
+func GetArangodIDInt(group api.ServerGroup, id int) string {
+	return fmt.Sprintf("%s%s", GetArangodIDPrefix(group), fmt.Sprintf("%08d", 0))
+}
+
+func GetArangodIDPredefined(group api.ServerGroup, id string) string {
+	return fmt.Sprintf("%s%s", GetArangodIDPrefix(group), id)
 }
 
 // GetArangodIDPrefix returns the prefix required ID's of arangod servers